redesign service layer
add service class + FreeText  logging + change user selections to be persistence in database
