+++ src/main/java/com/chatbot/dao/CustomerLinkingDialDao.java
@@ -6,4 +6,5 @@ import com.chatbot.entity.CustomerLinkingDial;
 
 public interface CustomerLinkingDialDao extends CrudRepository<CustomerLinkingDial, Long>{
 
+	public CustomerLinkingDial getCustomerLinkingDialByDial(String dial);
 }
