+++ src/main/java/com/chatbot/services/ChatBotServiceImpl.java
@@ -23,6 +23,7 @@ import com.chatbot.dao.FreeTextLoggingDao;
 import com.chatbot.dao.InteractionLoggingRepo;
 import com.chatbot.dao.InteractionMessageRepo;
 import com.chatbot.dao.PersistenceMenuButtonRepo;
+import com.chatbot.dao.UserSelectionDao;
 import com.chatbot.entity.BotButton;
 import com.chatbot.entity.BotButtonTemplateMSG;
 import com.chatbot.entity.BotConfiguration;
@@ -41,6 +42,7 @@ import com.chatbot.entity.EnabledCategoryConfiguration;
 import com.chatbot.entity.FreeTextLogging;
 import com.chatbot.entity.InteractionLogging;
 import com.chatbot.entity.PersistenceMenuButton;
+import com.chatbot.entity.UserSelection;
 
 @Service
 public class ChatBotServiceImpl implements ChatBotService {
@@ -53,6 +55,9 @@ public class ChatBotServiceImpl implements ChatBotService {
 	@Autowired
 	private BotButtonRepo botButtonRepo;
 
+	@Autowired
+	private UserSelectionDao userSelectionDao;
+	
 	@Autowired
 	private BotInteractionRepo botInteractionRepo;
 
@@ -261,6 +266,22 @@ public class ChatBotServiceImpl implements ChatBotService {
 	@Override
 	public CustomerLinkingDial saveCustomerLinkingDial(CustomerLinkingDial customerLinkingDial) {
 		return customerLinkingDialDao.save(customerLinkingDial);
+	}
+
+	@Override
+	public CustomerLinkingDial getCustomerLinkingDialById(String dial) {
+		return customerLinkingDialDao.getCustomerLinkingDialByDial(dial);
+	}
+
+	@Override
+	public UserSelection getUserSelectionBySenderId(String senderId) {
+		return userSelectionDao.findBySenderId(senderId);
+	}
+
+	@Override
+	public void updateUserSelections(UserSelection userSelection) {
+		userSelectionDao.save(userSelection);
+		
 	}
