+++ src/main/resources/application.properties
@@ -13,14 +13,14 @@ logging.file=/logs/social/chatbot
 logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} - %msg%n
 logging.pattern.file=%d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n
 
-spring.datasource.jndi-name=jdbc/oracledb
+#spring.datasource.jndi-name=jdbc/oracledb
 
 #Basic Spring Boot Config for Oracle Locally 
-#spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
-#spring.datasource.url=jdbc:oracle:thin:@//localhost:1521/xe
+spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
+spring.datasource.url=jdbc:oracle:thin:@//localhost:1521/xe
 #jdbc:oracle:thin:@//10.195.2.153:1522/xe
-#spring.datasource.username=messengerChatBot
-#spring.datasource.password=chatbot
+spring.datasource.username=messengerChatBot
+spring.datasource.password=chatbot
 
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
