+++ src/main/webapp/WEB-INF/weblogic.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="UTF-8"?>
+<!-- <?xml version="1.0" encoding="UTF-8"?>
 <wls:weblogic-web-app
 	xmlns:wls="http://xmlns.oracle.com/weblogic/weblogic-web-app"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
@@ -17,4 +17,4 @@
 			<java-charset-name>ISO8859_6</java-charset-name>
 		</input-charset>
 	</charset-params>
-</wls:weblogic-web-app>
+</wls:weblogic-web-app> -->
